Create simple-login-model.js
This login page handles form submission using the handleSubmit function, which makes an API call to verify the user's credentials. If the login is successful, the user is redirected to the homepage using the useRouter hook. If there is an error, the error message is displayed on the page. The useState hook is used to manage the state of the email, password, and error message fields in the form.