Create list-manage-model.js
This code creates a simple list with an input field to add new items and a button to remove items. The useState hook is used to manage the state of the list items and the text input for adding new items. The addItem function adds a new item to the list when the user clicks the "Add" button, and the removeItem function removes an item from the list when the user clicks the "Remove" button next to the item. The list is rendered using the map method on the listItems array.